+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Fixes
+* [maven-plugins] 146 / Pull operations fail on Windows with release 2.34.0
+
 ## v2.34.0 - 05 Jul 2023
 
 ### Enhancements